requirements.txt
internetarchive>=3.5.0
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
openai>=1.3.0
requests>=2.31.0
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
tqdm>=4.66.0
flask>=3.0.0
flask-cors>=4.0.0
flask-login>=0.6.3
werkzeug>=3.0.0
gunicorn>=21.2.0
librosa>=0.10.0
soundfile>=0.12.1
numba>=0.58.0
